% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{aeh_spatial_patterns}
\alias{aeh_spatial_patterns}
\title{Automatic Expression Histology (AEH)}
\usage{
aeh_spatial_patterns(X, exp_mat, DE_mll_results, C, l, ...)
}
\arguments{
\item{X}{Data frame or matrix with spatial coordinates.}

\item{exp_mat}{Data frame or matrix with gene expression at each cell/loocation, with
genes in columns. The data is assumed to be properly normalized.}

\item{DE_mll_results}{The results from SpatialDE, only with rows for statistically
significant genes.}

\item{C}{Number of expected patterns.}

\item{l}{Length scale for the patterns.}

\item{...}{Extra arguments passed to the function \code{fit_patterns}.
See \href{https://github.com/Teichlab/SpatialDE/blob/23ec19131473bf87d5fb3b17508aae430f2e6310/Python-module/SpatialDE/aeh.py#L137}{the Python module this package wraps}
for the available arguments.}
}
\value{
A list of 2 data frames, one with pattern membership information
for each gene, and the other with the posterior mean underlying expression for genes in
given spatial patterns.
}
\description{
Classify the patterns of spatially differentially expressed genes.
}
