% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{AEHSpatialPatterns}
\alias{AEHSpatialPatterns}
\alias{AEHSpatialPatterns.default}
\alias{AEHSpatialPatterns.SingleCellExperiment}
\alias{AEHSpatialPatterns.seurat}
\title{Automatic Expression Histology (AEH)}
\usage{
AEHSpatialPatterns(exp_mat, ...)

\method{AEHSpatialPatterns}{default}(exp_mat, X, DE_mll_results, C, l,
  row_sample = FALSE, ...)

\method{AEHSpatialPatterns}{SingleCellExperiment}(exp_mat, DE_mll_results,
  C, l, covariates = NULL, location_names = c("x", "y"),
  kernel_space = NULL, ...)

\method{AEHSpatialPatterns}{seurat}(exp_mat, DE_mll_results, C, l,
  covariates = NULL, location_names = c("x", "y"),
  kernel_space = NULL, ...)
}
\arguments{
\item{exp_mat}{Data frame or matrix with gene expression at each cell/loocation.
The data is assumed to be properly normalized. Genes can be in either columns or
rows. If the genes are in columns (and samples/cells in rows), set \code{row_sample = TRUE}.
Alternatively, a \code{SingleCellExperiment} or a \code{seurat} object can be used.
If it's a \code{SingleCellExperiment} object, it must have spatial coordinates in
\code{colData}. If it's a \code{seurat} object, it must have spatial coordinates
in the \code{meta.data} slot.}

\item{...}{Extra arguments passed to the function \code{fit_patterns}.
See \href{https://github.com/Teichlab/SpatialDE/blob/23ec19131473bf87d5fb3b17508aae430f2e6310/Python-module/SpatialDE/aeh.py#L137}{the Python module this package wraps}
for the available arguments.}

\item{X}{Data frame or matrix with spatial coordinates. All columns must be numeric,
referring to spatial coordinates. This argument is only needed when \code{exp_mat}
is a matrix or data frame.}

\item{DE_mll_results}{The results from SpatialDE, only with rows for statistically
significant genes.}

\item{C}{Number of expected patterns.}

\item{l}{Length scale for the patterns.}

\item{row_sample}{Whether samples/cells are in rows and genes are in columns. Defaults
to \code{FALSE}. This argument is only needed when \code{exp_mat}
is a matrix or data frame.}

\item{covariates}{Covariates that are not genes to which the test is also run,
as a comparison to genes. For example, log total counts in cells. This should
be a character vector, as column names of \code{colData} in \code{SingleCellExperiment}
objects, or column names of \code{meta.data} in \code{seurat} objects.}

\item{location_names}{Column names of column data (for \code{SingleCellExperiment}) or names of
columns of \code{meta.data} (for \code{seurat}) that specify the spatial coordinates
of cells.}

\item{kernel_space}{The grid of covariance matrices to search over for the alternative
model can be specified using the kernel_space paramter. This should be a named list,
each name specifying a kernel, and each element a numeric vector specifying
different parameters for that kernel to be tried.}
}
\value{
A list of 2 data frames, one with pattern membership information
for each gene, and the other with the posterior mean underlying expression for genes in
given spatial patterns.
}
\description{
Classify the patterns of spatially differentially expressed genes.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: AEH with matrix or data frame input

\item \code{SingleCellExperiment}: AEH with \code{SingleCellExperiment} input

\item \code{seurat}: AEH with \code{seurat} input
}}

