% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert2df}
\alias{convert2df}
\title{Convert inputs into the correct format}
\usage{
convert2df(X, exp_mat, row_sample = FALSE)
}
\arguments{
\item{X}{Data frame or matrix with spatial coordinates. All columns must be numeric,
referring to spatial coordinates.}

\item{exp_mat}{Data frame or matrix with gene expression at each cell/loocation.
The data is assumed to be properly normalized. Genes can be in either columns or
rows. If the genes are in columns (and samples/cells in rows), set \code{row_sample = TRUE}.}

\item{row_sample}{Whether samples/cells are in rows and genes are in columns. Defaults
to \code{FALSE}.}
}
\value{
A list of two data frames
}
\description{
The Python module takes in data frames as input, but you may pass a matrix to the
wrapper functions. This function converts whatever passed to the wrapper functions
into data frames.
}
