% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.importPyModule}
\alias{.importPyModule}
\title{Import SpatialDE}
\usage{
.importPyModule(proc, patch_tqdm, set_fake_tqdm, set_real_tqdm)
}
\arguments{
\item{patch_tqdm}{If \code{TRUE} patch calls to tqdm.}
}
\value{
An R wrapper for the SpatialDE Python module.
}
\description{
This function loads the SpatialDE Python module and optionally
monkey-patches it to remove tqdm calls.
}
