% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveDE.R
\name{.naiveDE_regress_out}
\alias{.naiveDE_regress_out}
\title{Wrapper for NaiveDE.regress_out python function}
\usage{
.naiveDE_regress_out(sample_info, dfm)
}
\arguments{
\item{sample_info}{dataframe with samples as rows,
'x', 'y' coordinates and  total raw counts as columns.}

\item{dfm}{dataframe resulting from naiveDE_stabilize}
}
\description{
Wrapper for NaiveDE.regress_out python function
}
