% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{FSV_sig}
\alias{FSV_sig}
\title{Plot Fraction Spatial Variance vs Q-value}
\usage{
FSV_sig(
  results,
  ms_results = NULL,
  certain_only = FALSE,
  log_x = FALSE,
  do_label = TRUE,
  covariate_names = NULL
)
}
\arguments{
\item{results}{Results from SpatialDE.}

\item{ms_results}{Model selection results, should be a data frame with columns
\code{g} for gene names and \code{model} for the model selected.}

\item{certain_only}{Only plot results with narrow 95\\% confidence interval.}

\item{log_x}{Whether to display x axis in log scale.}

\item{do_label}{Display gene names for statistically significant genes, default to \code{TRUE}.}

\item{covariate_names}{Names of covariates as a reference, default to \code{NULL}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Optionally provide model selection results to the function will color points by model.
Point size corresponds to certinety of the FSV value.
}
