% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{list_spatialDE_methods}
\alias{list_spatialDE_methods}
\title{List available methods from SpatialDE}
\usage{
list_spatialDE_methods()
}
\description{
Shows the available methods from the
\href{https://github.com/Teichlab/SpatialDE}{SpatialDE} Python module.
}
\examples{
list_spatialDE_methods()
}
