% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{locations}
\alias{locations}
\title{Location of Spots in the Mouse Olfactory Bulb Data}
\format{A data frame with the following columns:
\describe{
  \item{X1}{Spots names}
  \item{x}{x coordinate of the location of the spots}
  \item{y}{y coordinate of the location of the spots}
}}
\source{
\url{https://doi.org/10.1126/science.aaf2403}
}
\usage{
locations
}
\description{
Location of Spots in the Mouse Olfactory Bulb Data
}
\keyword{datasets}
