% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOB_sample_info}
\alias{MOB_sample_info}
\title{Mouse Olfactory Bulb sample metadata}
\format{
A \code{data.frame} with 262 rows and 3 variables as columns: the \code{x} and
\code{y} coordinates and \code{total_counts} corresponding to each spot.
}
\usage{
data(MOB_sample_info)
}
\description{
Coordinates and total counts for the samples from the Mouse Olfactory Bulb
data generated by Stahl et al. (2016).
This data was originally downloaded from
\url{https://github.com/Teichlab/SpatialDE/blob/master/Analysis/MouseOB/MOB_sample_info.csv}.
}
\references{
Ståhl, P. L. et al. (2016) 'Visualization and analysis of gene expression in
tissue sections by spatial transcriptomics', \emph{Science}, 353(6294), p.
78. doi: 10.1126/science.aaf2403.
}
\keyword{datasets}
