% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mockSVG}
\alias{mockSVG}
\title{Generate count matrix for spatially variable genes.}
\usage{
mockSVG(size, tot_genes, de_genes)
}
\arguments{
\item{size}{\code{int} genes will be spatially arranged on a size x size grid}

\item{tot_genes}{\code{int} total number of genes}

\item{de_genes}{\code{int} number of spatially variable genes}
}
\value{
\code{list} containing:
\itemize{
\item \code{coordinates}: \code{data.frame} with \code{x} and \code{y} columns;
\item \code{counts}: \code{matrix} with generated gene counts.
}
}
\description{
Generate count matrix for spatially variable genes.
}
