% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{model_search}
\alias{model_search}
\title{Compare model fits with different models}
\usage{
model_search(x, coordinates, de_results, verbose = FALSE)
}
\arguments{
\item{x}{\code{matrix}-like object of normalized counts. E.g. resulting from
\code{\link[=regress_out]{regress_out()}}.}

\item{coordinates}{\code{data.frame} with sample coordinates.
Each row is a sample, the columns with coordinates must be named 'x' and 'y'.}

\item{de_results}{\code{data.frame} resulting from \code{\link[=run]{run()}} filtered
based on \code{qvalue < threshold} (e.g. \code{qvalue < 0.05})}

\item{verbose}{\code{logical} controlling the display of the progress bar.}
}
\value{
\code{data.frame} of model_search results.
}
\description{
Classify DE genes to interpretable fitting classes.
}
\examples{
set.seed(42)
mock <- mockSVG(size = 10, tot_genes = 300, de_genes = 10)
stabilized <- stabilize(mock$counts)
samples_info <- mock$coordinates
samples_info$total_counts <- colSums(mock$counts)
regressed <- regress_out(samples_info, stabilized)

## Run SpatialDE
de_results <- run(counts = regressed, coordinates = mock$coordinates)

## Run model search
ms_results <- model_search(mock$coordinates, regressed, de_results)
ms_results <- model_search(
    x = regressed,
    coordinates = mock$coordinates,
    de_results = de_results
)

}
