% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mouseob}
\alias{mouseob}
\title{Mouse Olfactory Bulb Data from Stahl et al.}
\format{A matrix with 1000 columns (genes) and 260 rows (spots), with the gene
symbols in the column name and spot names in the row name.}
\source{
\url{https://doi.org/10.1126/science.aaf2403}
}
\usage{
mouseob
}
\description{
This is a sample of the mouse olfactory bulb data from
\href{https://doi.org/10.1126/science.aaf2403}{Stahl et al}.
Out of the over 16000 genes in the original data set, 1000 genes were chosen at
random. We used the Rep11 data. See
\href{http://www.spatialtranscriptomicsresearch.org/datasets/doi-10-1126science-aaf2403/}{this site}
for the raw data.
}
\keyword{datasets}
