% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mouseob_small}
\alias{mouseob_small}
\title{Smaller Subset of Mouse Olfactory Bulb Data}
\format{A matrix with 100 columns (genes) and 260 rows (spots), with the gene
symbols in the column name and spot names in the row name.}
\source{
\url{https://doi.org/10.1126/science.aaf2403}
}
\usage{
mouseob_small
}
\description{
This is a smaller subset of the mouse olfactory bulb data from
\href{https://doi.org/10.1126/science.aaf2403}{Stahl et al}, with only 100 randomly
selected genes. This dataset is only used for examples and unit testing.
}
\keyword{datasets}
