% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{multiGenePlot}
\alias{multiGenePlot}
\title{Plot Spatial Patterns of Multiple Genes}
\usage{
multiGenePlot(
  normalized_counts,
  coordinates,
  genes_plot,
  viridis_option = "D",
  ncol = 2,
  point_size = 1,
  dark_theme = TRUE
)
}
\arguments{
\item{normalized_counts}{\code{matrix} of variance stabilized counts,
e.g. resulting from \code{\link[=stabilize]{stabilize()}}.}

\item{coordinates}{\code{data.frame} with sample coordinates.
Each row is a sample, the columns names must be named 'x' and 'y'.}

\item{genes_plot}{character vector specifying which genes are to be plotted.}

\item{viridis_option}{This function uses the \code{viridis} palette to color
cells for gene expression. Four options are available: "magma" (or "A"),
"inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the default option)
and "cividis" (or "E").}

\item{ncol}{Number of columns to arrange the plots.}

\item{point_size}{Point size of each plot.}

\item{dark_theme}{Whether dark background should be used; this is helpful to
highlight cells with high expression when using the \code{viridis} palette.}
}
\value{
This function draws a plot for each specified genes
}
\description{
Plot Spatial Patterns of Multiple Genes
}
