% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveDE.R
\name{naiveDE_stabilize}
\alias{naiveDE_stabilize}
\title{Wrapper for SpatialDE.run python function}
\usage{
naiveDE_stabilize(df)
}
\arguments{
\item{df}{dataframe with the expression values. Columns are genes, and Rows
are samples}
}
\description{
Wrapped functions from NaiveDE python package
\href{https://github.com/Teichlab/NaiveDE}{NaiveDE}
}
