% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveDE.R
\name{regress_out}
\alias{regress_out}
\title{Wrapper for NaiveDE.regress_out python function}
\usage{
regress_out(sample_info, stabilized_counts)
}
\arguments{
\item{sample_info}{\code{data.frame} with samples as rows,
'x', 'y' coordinates and  total raw counts as columns.}

\item{stabilized_counts}{\code{matrix} or \code{data.frame} resulting from
naiveDE_stabilize}
}
\description{
Wrapper for NaiveDE.regress_out python function
}
