% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rep11_MOB_0}
\alias{Rep11_MOB_0}
\title{Mouse Olfactory Bulb spatial gene expression data}
\format{
A matrix with 16218 genes as rows and 262 spots as columns.
}
\usage{
data(Rep11_MOB_0)
}
\description{
Replicate 11 from the spatially dependent gene expression data from the mouse
olfactory bulb generated by Stahl et al. (2016).
This data was originally downloaded from
\url{https://github.com/Teichlab/SpatialDE/blob/master/Analysis/MouseOB/data/Rep11_MOB_0.csv}.
}
\references{
Ståhl, P. L. et al. (2016) 'Visualization and analysis of gene expression in
tissue sections by spatial transcriptomics', \emph{Science}, 353(6294), p.
78. doi: 10.1126/science.aaf2403.
}
\keyword{datasets}
