% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{run}
\alias{run}
\title{Perform SpatialDE test}
\usage{
run(x, coordinates, verbose = FALSE)
}
\arguments{
\item{x}{\code{matrix}-like object of normalized counts. E.g. resulting from
\code{\link[=regress_out]{regress_out()}}.}

\item{coordinates}{\code{data.frame} with sample coordinates.
Each row is a sample, the columns with coordinates must be named 'x' and 'y'.}

\item{verbose}{\code{logical} controlling the display of the progress bar.}
}
\value{
\code{data.frame} with DE results.
}
\description{
Wraps the \code{run} function from the
\href{https://github.com/Teichlab/SpatialDE}{SpatialDE} Python package.
}
\examples{
set.seed(42)
mock <- mockSVG(size = 10, tot_genes = 500, de_genes = 10)
stabilized <- stabilize(mock$counts)
sample_info <- mock$coordinates
sample_info$total_counts <- colSums(mock$counts)
regressed <- regress_out(counts = stabilized, sample_info = sample_info)

## Run SpatialDE
de_results <- run(counts = regressed, coordinates = mock$coordinates)

}
