% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{run_model_search}
\alias{run_model_search}
\title{Wrapper for SpatialDE.model_search python function}
\usage{
run_model_search(coordinates, regressed_counts, de_results)
}
\arguments{
\item{coordinates}{\code{data.frame} with sample coordinates.
Each row is a sample, the columns with coordinates must be named 'x' and 'y'.}

\item{regressed_counts}{\code{data.frame} resulting from \code{naiveDE_regress_out}}

\item{de_results}{\code{data.frame} resulting from \code{spatialDE_run} filtered based
on qvalue < threshold (e.g. qvalue < 0.05)}
}
\description{
Wrapper for SpatialDE.model_search python function
}
