% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{run_spatialDE}
\alias{run_spatialDE}
\title{Wrapper for SpatialDE.run python function}
\usage{
run_spatialDE(coordinates, regressed_counts)
}
\arguments{
\item{coordinates}{\code{data.frame} with sample coordinates.
Each row is a sample, the columns with coordinates must be named 'x' and 'y'.}

\item{regressed_counts}{\code{data.frame} or \code{matrix} resulting from
\code{naiveDE_regress_out}}
}
\description{
Wrapper for SpatialDE.run python function
}
\note{
both functions of SpatialDE run with tqdm python library, which
allow to view a progress bar and it's not suitable for R.
TODO: hide prints of progress bar
}
