% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{RunSpatialDE}
\alias{RunSpatialDE}
\alias{RunSpatialDE.default}
\alias{RunSpatialDE.SingleCellExperiment}
\alias{RunSpatialDE.seurat}
\title{Perform SpatialDE test}
\usage{
RunSpatialDE(exp_mat, ...)

\method{RunSpatialDE}{default}(exp_mat, X, kernel_space = NULL,
  row_sample = FALSE)

\method{RunSpatialDE}{SingleCellExperiment}(exp_mat, covariates = NULL,
  location_names = c("x", "y"), kernel_space = NULL)

\method{RunSpatialDE}{seurat}(exp_mat, covariates = NULL,
  location_names = c("x", "y"), kernel_space = NULL)
}
\arguments{
\item{exp_mat}{Data frame or matrix with gene expression at each cell/loocation.
The data is assumed to be properly normalized. Genes can be in either columns or
rows. If the genes are in columns (and samples/cells in rows), set \code{row_sample = TRUE}.
Alternatively, a \code{SingleCellExperiment} or a \code{seurat} object can be used.
If it's a \code{SingleCellExperiment} object, it must have spatial coordinates in
\code{colData}. If it's a \code{seurat} object, it must have spatial coordinates
in the \code{meta.data} slot.}

\item{X}{Data frame or matrix with spatial coordinates. All columns must be numeric,
referring to spatial coordinates. This argument is only needed when \code{exp_mat}
is a matrix or data frame.}

\item{kernel_space}{The grid of covariance matrices to search over for the alternative
model can be specified using the kernel_space paramter. This should be a named list,
each name specifying a kernel, and each element a numeric vector specifying
different parameters for that kernel to be tried.}

\item{row_sample}{Whether samples/cells are in rows and genes are in columns. Defaults
to \code{FALSE}. This argument is only needed when \code{exp_mat}
is a matrix or data frame.}

\item{covariates}{Covariates that are not genes to which the test is also run,
as a comparison to genes. For example, log total counts in cells. This should
be a character vector, as column names of \code{colData} in \code{SingleCellExperiment}
objects, or column names of \code{meta.data} in \code{seurat} objects.}

\item{location_names}{Column names of column data (for \code{SingleCellExperiment}) or names of
columns of \code{meta.data} (for \code{seurat}) that specify the spatial coordinates
of cells.}
}
\value{
A data frame with the following columns:
\describe{
\item{\code{FSV}}{Fraction of total variance explained by spatial variance}
\item{\code{M}}{Internal code for model used}
\item{\code{g}}{Gene names}
\item{\code{l}}{Length scale the gene changes expression over}
\item{\code{max_delta}}{Maximum likelihood estimate of ratio between variance
 of non-spatial variance
and spatial variance}
\item{\code{max_ll}}{Maximum log likelihood}
\item{\code{max_mu_hat}}{Mean expression level for the gene}
\item{\code{max_s2_t_hat}}{}
\item{\code{model}}{The model used to fit the data, which indicates spatial pattern
of gene expression}
\item{\code{n}}{}
\item{\code{s2_FSV}}{Standard error of FSV}
\item{\code{s2_logdelta}}{Standard error of log delta}
\item{\code{time}}{Time taken on this gene}
\item{\code{BIC}}{Akaike's An Information Criterion, with k = log(n); a smaller number
means the model fits the data better}
\item{\code{max_ll_null}}{Maximum likelihood of null model}
\item{\code{LLR}}{Log likelihood ratio}
\item{\code{pval}}{P-value of spatial differential expression}
\item{\code{qval}}{Significance after correcting for multiple testing}}
}
\description{
Test which gene is dependent on spatial location in the tissue. For more information,
see the original paper
\href{https://doi.org/10.1038/nmeth.4636}{SpatialDE: identification of spatially variable genes}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Perform SpatialDE test on matrices

\item \code{SingleCellExperiment}: Perform SpatialDE test on \code{SingleCellExperiment} objects

\item \code{seurat}: Perform SpatialDE test on \code{seurat} objects
}}

