% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sce2mat}
\alias{sce2mat}
\title{Convert \code{SingleCellExperiment} to matrix}
\usage{
sce2mat(object, covariates, location_names)
}
\arguments{
\item{object}{The \code{SingleCellExperiment} object.}

\item{covariates}{Covariates that are not genes to which the test is also run,
as a comparison to genes. For example, log total counts in cells. This should
be a character vector, as column names of \code{colData} in \code{SingleCellExperiment}
objects, or column names of \code{meta.data} in \code{Seurat} objects.}

\item{location_names}{Column names of column data (for \code{SingleCellExperiment}) or names of
columns of \code{meta.data} (for \code{Seurat}) that specify the spatial coordinates
of cells.}
}
\value{
A list with a matrix and a data frame
}
\description{
This function pulls out relevant parts of the \code{SingleCellExperiment} object
and converts it into matrix before calling the Python module `SpatialDE`.
}
