% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{spatial_patterns}
\alias{spatial_patterns}
\title{Group spatially variable genes into spatial patterns using automatic
expression histology (AEH)}
\usage{
spatial_patterns(
  coordinates,
  regressed_counts,
  de_results,
  C,
  l,
  verbose = FALSE
)
}
\arguments{
\item{coordinates}{\code{data.frame} with sample coordinates.
Each row is a sample, the columns with coordinates must be named 'x' and 'y'.}

\item{regressed_counts}{\code{matrix} resulting from \code{\link[=regress_out]{regress_out()}}.}

\item{de_results}{\code{data.frame} resulting from \code{\link[=run]{run()}} filtered
based on \code{qvalue < threshold} (e.g. \code{qvalue < 0.05})}

\item{C}{\code{integer} The number of spatial patterns}

\item{l}{\code{numeric} The charancteristic length scale of the clusters}

\item{verbose}{\code{bool} controlling the display of the progress messages.}
}
\value{
\code{list} of two dataframe (pattern_results, patterns):
\code{pattern_results} dataframe with pattern membership information for each
gene.
\code{patterns} the posterior mean underlying expression fro genes in given
spatial patterns.
}
\description{
Group spatially variable genes into spatial patterns using automatic
expression histology (AEH)
}
\examples{
set.seed(42)
mock <- mockSVG(10, 1000, 10)
stabilized <- stabilize(mock$counts)
samples_info <- mock$coordinates
samples_info$total_counts <- colSums(mock$counts)
regressed <- regress_out(samples_info, stabilized)

## Run SpatialDE
results <- run(mock$coordinates, regressed)
de_results <- results[results$qval<0.1, ]

## Run Spatial_patterns
sp <- spatial_patterns(coordinates, regressed, de_results = de_results, 
                       C = 5, l = 1.5)

sp$pattern_results
sp$patterns

}
