% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{spatialDE_model_search}
\alias{spatialDE_model_search}
\title{Wrapper for SpatialDE.model_search python function}
\usage{
spatialDE_model_search(coordinates, res, de_results)
}
\arguments{
\item{de_results}{dataframe resulting from spatialDE_run filtered based on
qvalue < threshold (e.g. qvalue < 0.05)}
}
\description{
Wrapper for SpatialDE.model_search python function
}
