% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{spatialDE_run}
\alias{spatialDE_run}
\title{Wrapper for SpatialDE.run python function}
\usage{
spatialDE_run(coordinates, res)
}
\arguments{
\item{coordinates}{dataframe with sample coordinates. Each row is a sample,
the columns with coordinates must be named 'x' and 'y'.}

\item{res}{dataframe resulting from naiveDE_regress_out function}
}
\description{
Wrapper for SpatialDE.run python function
}
\note{
both functions of SpatialDE run with tqdm python library, which
allow to view a progress bar and it's not suitable for R.
TODO: hide prints of progress bar
}
