% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveDE.R
\name{stabilize}
\alias{stabilize}
\title{Wrapper for NaiveDE.stabilize python function}
\usage{
stabilize(counts)
}
\arguments{
\item{counts}{\code{matrix} or \code{data.frame} with the expression values.
Columns should be samples, and Rows genes}
}
\description{
Wrapped functions from NaiveDE python package
\href{https://github.com/Teichlab/NaiveDE}{NaiveDE}
}
