% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{model_search}
\alias{model_search}
\title{Compare model fits with different models}
\usage{
model_search(coordinates, regressed_counts, de_results, verbose = FALSE)
}
\arguments{
\item{coordinates}{\code{data.frame} with sample coordinates.
Each row is a sample, the columns with coordinates must be named 'x' and 'y'.}

\item{regressed_counts}{\code{matrix} resulting from \code{\link[=regress_out]{regress_out()}}}

\item{de_results}{\code{data.frame} resulting from \code{\link[=run]{run()}} filtered
based on \code{qvalue < threshold} (e.g. \code{qvalue < 0.05})}

\item{verbose}{\code{bool} controlling the display of the progress bar.}
}
\value{
\code{data.frame} of model_search results.
}
\description{
Classify DE genes to interpretable fitting classes.
}
\examples{
\dontrun{
ncells <- 100
ngenes <- 1000
counts <- matrix(rpois(ncells * ngenes, lambda = 3),
                 nrow = ngenes, ncol = ncells)

## Provide total counts for library size normalization and coordinates
sample_info <- data.frame(x = rnorm(ncells), y = rnorm(ncells), 
                          total_counts = colSums(counts)) 
coordinates <- sample_info[, c("x", "y")]

stabilized <- stabilize(counts)
regressed <- regress_out(sample_info, stabilized)

## Run SpatialDE
de_results <- run(coordinates, regressed)

## Run model search
ms_results <- model_search(coordinates, regressed, de_results)
}
}
