% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{ModelSearch}
\alias{ModelSearch}
\alias{ModelSearch.default}
\alias{ModelSearch.SingleCellExperiment}
\alias{ModelSearch.seurat}
\title{Compare Model Fits with Different Models}
\usage{
ModelSearch(exp_mat, ...)

\method{ModelSearch}{default}(exp_mat, X, DE_mll_results,
  kernel_space = NULL, row_sample = FALSE)

\method{ModelSearch}{SingleCellExperiment}(exp_mat, DE_mll_results,
  covariates = NULL, location_names = c("x", "y"),
  kernel_space = NULL)

\method{ModelSearch}{seurat}(exp_mat, DE_mll_results, covariates = NULL,
  location_names = c("x", "y"), kernel_space = NULL)
}
\arguments{
\item{exp_mat}{Data frame or matrix with gene expression at each cell/loocation.
The data is assumed to be properly normalized. Genes can be in either columns or
rows. If the genes are in columns (and samples/cells in rows), set \code{row_sample = TRUE}.
Alternatively, a \code{SingleCellExperiment} or a \code{seurat} object can be used.
If it's a \code{SingleCellExperiment} object, it must have spatial coordinates in
\code{colData}. If it's a \code{seurat} object, it must have spatial coordinates
in the \code{meta.data} slot.}

\item{X}{Data frame or matrix with spatial coordinates. All columns must be numeric,
referring to spatial coordinates. This argument is only needed when \code{exp_mat}
is a matrix or data frame.}

\item{DE_mll_results}{The results from SpatialDE, only with rows for statistically
significant genes.}

\item{kernel_space}{The grid of covariance matrices to search over for the alternative
model can be specified using the kernel_space paramter. This should be a named list,
each name specifying a kernel, and each element a numeric vector specifying
different parameters for that kernel to be tried.}

\item{row_sample}{Whether samples/cells are in rows and genes are in columns. Defaults
to \code{FALSE}. This argument is only needed when \code{exp_mat}
is a matrix or data frame.}

\item{covariates}{Covariates that are not genes to which the test is also run,
as a comparison to genes. For example, log total counts in cells. This should
be a character vector, as column names of \code{colData} in \code{SingleCellExperiment}
objects, or column names of \code{meta.data} in \code{seurat} objects.}

\item{location_names}{Column names of column data (for \code{SingleCellExperiment}) or names of
columns of \code{meta.data} (for \code{seurat}) that specify the spatial coordinates
of cells.}
}
\value{
A data frame
}
\description{
This way DE genes can be classified to interpretable function classes.
The strategy is based on ABCD in the Automatic Statistician, but
using precomputed covariance matrices for all models in the search space.
By default searches a grid of periodic covariance matrices and a linear
covariance matrix.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Compare models with matrix or data frame input

\item \code{SingleCellExperiment}: Compare models with \code{SingleCellExperiment} input

\item \code{seurat}: Compare models with \code{seurat} input
}}

