% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{MultiGenePlot}
\alias{MultiGenePlot}
\alias{MultiGenePlot.default}
\alias{MultiGenePlot.SingleCellExperiment}
\alias{MultiGenePlot.seurat}
\title{Plot Spatial Patterns of Multiple Genes}
\usage{
MultiGenePlot(exp_mat, ...)

\method{MultiGenePlot}{default}(exp_mat, X, genes_plot,
  row_sample = FALSE, facet_titles = NULL, viridis_option = "D",
  ncol = 2, dark_theme = TRUE)

\method{MultiGenePlot}{SingleCellExperiment}(exp_mat, genes_plot,
  location_names = c("x", "y"), facet_titles = NULL,
  viridis_option = "D", ncol = 2, dark_theme = TRUE)

\method{MultiGenePlot}{seurat}(exp_mat, genes_plot,
  location_names = c("x", "y"), facet_titles = NULL,
  viridis_option = "D", ncol = 2, dark_theme = TRUE)
}
\arguments{
\item{exp_mat}{Data frame or matrix with gene expression at each cell/loocation.
The data is assumed to be properly normalized. Genes can be in either columns or
rows. If the genes are in columns (and samples/cells in rows), set \code{row_sample = TRUE}.
Alternatively, a \code{SingleCellExperiment} or a \code{seurat} object can be used.
If it's a \code{SingleCellExperiment} object, it must have spatial coordinates in
\code{colData}. If it's a \code{seurat} object, it must have spatial coordinates
in the \code{meta.data} slot.}

\item{X}{Data frame or matrix with spatial coordinates. All columns must be numeric,
referring to spatial coordinates. This argument is only needed when \code{exp_mat}
is a matrix or data frame.}

\item{genes_plot}{Character vector specifying which genes are to be plotted.
Covariate names are permitted in this argument.}

\item{row_sample}{Whether samples/cells are in rows and genes are in columns. Defaults
to \code{FALSE}. This argument is only needed when \code{exp_mat}
is a matrix or data frame.}

\item{facet_titles}{Title for each facet. This must be a character vector with the
same length as \code{genes_plot}.}

\item{viridis_option}{This function uses the \code{viridis} palette to color cells
for gene expression. Four options are available: "magma" (or "A"), "inferno" (or "B"),
"plasma" (or "C"), "viridis" (or "D", the default option) and "cividis" (or "E").}

\item{ncol}{Number of columns to arrange the plots.}

\item{dark_theme}{Whether dark background should be used; this is helpful to highlight
cells with high expression when using the \code{viridis} palette.}

\item{location_names}{Column names of column data (for \code{SingleCellExperiment}) or names of
columns of \code{meta.data} (for \code{seurat}) that specify the spatial coordinates
of cells.}
}
\value{
This function draws on the current divice without returning an object.
}
\description{
While \code{ggplot2} allows us to easily generate multi-faceted plots, we can't
use a separate color scale for each facet. This function allows you to make
multi-faceted plots with a separate color scale for each facet, so high expression
of one gene in the facets will not affect the dynamic range of colors for the other,
not so highly expressed genes, making it easier to discern spatial patterns.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Plot multiple genes for matrix or data frame input

\item \code{SingleCellExperiment}: Plot multiple genes for \code{SingleCellExperiment}
objects

\item \code{seurat}: Plot multiple genes for \code{seurat} objects
}}

