% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveDE.R
\name{regress_out}
\alias{regress_out}
\title{Regress out library size effect}
\usage{
regress_out(sample_info, stabilized_counts)
}
\arguments{
\item{sample_info}{\code{data.frame} with samples as rows and at least a column
with \code{total_counts}.}

\item{stabilized_counts}{\code{matrix} of variance stabilized
counts, e.g. resulting from \code{\link[=stabilize]{stabilize()}}.}
}
\value{
\code{matrix} of normalized counts.
}
\description{
Regresses out the effect of library size.
This function is a wrapper for \code{regress_out} from the
\href{https://github.com/Teichlab/NaiveDE}{NaiveDE} Python package.
}
\examples{
ncells <- 100
ngenes <- 1000
counts <- matrix(rpois(ncells * ngenes, lambda = 3),
                 nrow = ngenes, ncol = ncells)

## Provide total counts for library size normalization
sample_info <- data.frame(total_counts = colSums(counts))

stabilized <- stabilize(counts)
regressed <- regress_out(sample_info, stabilized)

}
