% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run}
\alias{run}
\title{Perform SpatialDE test}
\usage{
run(X, exp_mat, kernel_space = NULL)
}
\arguments{
\item{X}{Data frame or matrix with spatial coordinates.}

\item{exp_mat}{Data frame or matrix with gene expression at each cell/loocation, with
genes in columns. The data is assumed to be properly normalized.}

\item{kernel_space}{The grid of covariance matrices to search over for the alternative
model can be specified using the kernel_space paramter. This should be a named list,
each name specifying a kernel, and each element a numeric vector specifying
different parameters for that kernel to be tried.}
}
\value{
A data frame with the following columns:
\describe{
\item{\code{FSV}}{Fraction of total variance explained by spatial variance}
\item{\code{M}}{Internal code for model used}
\item{\code{g}}{Gene names}
\item{\code{l}}{Length scale the gene changes expression over}
\item{\code{max_delta}}{Maximum likelihood estimate of ratio between variance of non-spatial variance
and spatial variance}
\item{\code{max_ll}}{Maximum log likelihood}
\item{\code{max_mu_hat}}{Mean expression level for the gene}
\item{\code{max_s2_t_hat}}{}
\item{\code{model}}{The model used to fit the data, which indicates spatial pattern
of gene expression}
\item{\code{n}}{}
\item{\code{s2_FSV}}{Standard error of FSV}
\item{\code{s2_logdelta}}{Standard error of log delta}
\item{\code{time}}{Time taken on this gene}
\item{\code{BIC}}{Akaike's An Information Criterion, with k = log(n); a smaller number
means the model fits the data better}
\item{\code{max_ll_null}}{Maximum likelihood of null model}
\item{\code{LLR}}{Log likelihood ratio}
\item{\code{pval}}{P-value of spatial differential expression}
\item{\code{qval}}{Significance after correcting for multiple testing}}
}
\description{
Test which gene is dependent on spatial location in the tissue. For more information,
see the original paper
\href{https://doi.org/10.1038/nmeth.4636}{SpatialDE: identification of spatially variable genes}.
}
