% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{run}
\alias{run}
\title{Perform SpatialDE test}
\usage{
run(coordinates, regressed_counts, verbose = FALSE)
}
\arguments{
\item{coordinates}{\code{data.frame} with sample coordinates.
Each row is a sample, the columns with coordinates must be named 'x' and 'y'.}

\item{regressed_counts}{\code{matrix} resulting from \code{\link[=regress_out]{regress_out()}}.}

\item{verbose}{\code{bool} controlling the display of the progress bar.}
}
\value{
\code{data.frame} with DE results.
}
\description{
Wraps the \code{run} function from the
\href{https://github.com/Teichlab/SpatialDE}{SpatialDE} Python package.
}
\examples{
set.seed(42)
mock <- mockSVG(10, 1000, 10)
stabilized <- stabilize(mock$counts)
samples_info <- mock$coordinates
samples_info$total_counts <- colSums(mock$counts)
regressed <- regress_out(samples_info, stabilized)

## Run SpatialDE
de_results <- run(mock$coordinates, regressed)

}
