% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{run_spatialDE}
\alias{run_spatialDE}
\title{Perform SpatialDE test}
\usage{
run_spatialDE(coordinates, regressed_counts)
}
\arguments{
\item{coordinates}{\code{data.frame} with sample coordinates.
Each row is a sample, the columns with coordinates must be named 'x' and 'y'.}

\item{regressed_counts}{\code{data.frame} or \code{matrix} resulting from
\code{\link[=regress_out]{regress_out()}}.}
}
\value{
\code{data.frame} with DE results.
}
\description{
Wraps the \code{run} function from the
\href{https://github.com/Teichlab/SpatialDE}{SpatialDE} Python package.
}
\examples{
ncells <- 100
ngenes <- 1000
counts <- matrix(rpois(ncells * ngenes, lambda = 3),
                 nrow = ngenes, ncol = ncells)

## Provide total counts for library size normalization and coordinates
sample_info <- data.frame(total_counts = colSums(counts))
coordinates <- data.frame(x = rnorm(ncells), y = rnorm(ncells))

stabilized <- stabilize(counts)
regressed <- regress_out(sample_info, stabilized)

## Run SpatialDE
de_results <- run_spatialDE(coordinates, regressed)

}
