% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{spatial_patterns}
\alias{spatial_patterns}
\title{Group spatially variable genes into spatial patterns using automatic
expression histology (AEH)}
\usage{
spatial_patterns(
  x,
  coordinates,
  de_results,
  n_patterns,
  length,
  verbose = FALSE
)
}
\arguments{
\item{x}{\code{matrix}-like object of normalized counts. E.g. resulting from
\code{\link[=regress_out]{regress_out()}}.}

\item{coordinates}{\code{data.frame} with sample coordinates.
Each row is a sample, the columns with coordinates must be named 'x' and 'y'.}

\item{de_results}{\code{data.frame} resulting from \code{\link[=run]{run()}} filtered
based on \code{qvalue < threshold} (e.g. \code{qvalue < 0.05})}

\item{n_patterns}{\code{integer} The number of spatial patterns}

\item{length}{\code{numeric} The characteristic length scale of the clusters}

\item{verbose}{\code{logical} controlling the display of the progress bar.}
}
\value{
\code{list} of two dataframe (pattern_results, patterns):
\code{pattern_results} dataframe with pattern membership information for each
gene.
\code{patterns} the posterior mean underlying expression fro genes in given
spatial patterns.
}
\description{
Group spatially variable genes into spatial patterns using automatic
expression histology (AEH)
}
\examples{
set.seed(42)
mock <- mockSVG(size = 10, tot_genes = 500, de_genes = 10)
stabilized <- stabilize(mock$counts)
sample_info <- mock$coordinates
sample_info$total_counts <- colSums(mock$counts)
regressed <- regress_out(counts = stabilized, sample_info = sample_info)

## Run SpatialDE
results <- run(x = regressed, coordinates = mock$coordinates)
de_results <- results[results$qval<0.1, ]

## Run Spatial_patterns
sp <- spatial_patterns(
    x = x,
    coordinates = mock$coordinates,
    de_results = de_results,
    n_patterns = 5, length = 1.5
)

sp$pattern_results
sp$patterns

}
